Move unit tests into parent package
