+++ use-case-shopping/src/test/java/ai/vespa/example/shopping/site/SimpleHtmlBuilderTest.java
@@ -1,7 +1,8 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package ai.vespa.example.shopping.site.view;
+package ai.vespa.example.shopping.site;
 
+import ai.vespa.example.shopping.site.view.SimpleHtmlBuilder;
 import org.junit.Test;
 
 import java.util.Map;
