+++ use-case-shopping/src/test/java/ai/vespa/example/shopping/site/SimpleQueryBuilderTest.java
@@ -1,6 +1,6 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package ai.vespa.example.shopping.site.view;
+package ai.vespa.example.shopping.site;
 
 import ai.vespa.example.shopping.site.data.SimpleQueryBuilder;
 import org.junit.Test;
