+++ use-case-shopping/src/test/java/ai/vespa/example/shopping/site/SimpleTemplateTest.java
@@ -1,10 +1,11 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-package ai.vespa.example.shopping.site.view;
+package ai.vespa.example.shopping.site;
 
 import java.io.ByteArrayOutputStream;
 import java.nio.charset.StandardCharsets;
 
+import ai.vespa.example.shopping.site.view.SimpleTemplate;
 import org.junit.Test;
 
 import static junit.framework.TestCase.assertEquals;
